requirements.txt
requests>=2.28.0
python-docx>=1.0.0
pyyaml>=6.0
streamlit>=1.28.0
mcp[cli]>=1.0.0
